requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.25.2
fastapi==0.94.0
ffmpeg-python==0.2.0
filelock==3.9.0
future==0.18.3
h11==0.14.0
huggingface-hub==0.13.1
idna==3.4
lit==15.0.7
llvmlite==0.39.1
more-itertools==9.1.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai-whisper==20230308
packaging==23.0
pydantic==1.10.6
python-multipart==0.0.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sniffio==1.3.0
starlette==0.26.0.post1
tokenizers==0.13.2
torch==1.13.1
tqdm==4.65.0
transformers==4.26.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.21.0
fastapi==0.94.0
